move core models to `commonMain`
