+++ app/src/androidTest/kotlin/com/github/odaridavid/weatherapp/SettingsRepositoryTest.kt
@@ -5,12 +5,12 @@ import androidx.test.core.app.ApplicationProvider
 import androidx.test.filters.SmallTest
 import app.cash.turbine.test
 import com.github.odaridavid.weatherapp.core.api.SettingsRepository
-import com.github.odaridavid.weatherapp.core.model.DefaultLocation
-import com.github.odaridavid.weatherapp.core.model.ExcludedData
-import com.github.odaridavid.weatherapp.core.model.SupportedLanguage
-import com.github.odaridavid.weatherapp.core.model.TimeFormat
-import com.github.odaridavid.weatherapp.core.model.Units
 import com.github.odaridavid.weatherapp.data.settings.DefaultSettingsRepository
+import com.github.odaridavid.weatherapp.model.DefaultLocation
+import com.github.odaridavid.weatherapp.model.ExcludedData
+import com.github.odaridavid.weatherapp.model.SupportedLanguage
+import com.github.odaridavid.weatherapp.model.TimeFormat
+import com.github.odaridavid.weatherapp.model.Units
 import kotlinx.coroutines.test.runTest
 import org.junit.Before
 import org.junit.Test
