+++ app/src/main/java/com/github/odaridavid/weatherapp/core/api/WeatherRepository.kt
@@ -1,9 +1,8 @@
 package com.github.odaridavid.weatherapp.core.api
 
-import com.github.odaridavid.weatherapp.core.model.DefaultLocation
-import com.github.odaridavid.weatherapp.core.model.Weather
-import com.github.odaridavid.weatherapp.core.Result
-import kotlinx.coroutines.flow.Flow
+import com.github.odaridavid.weatherapp.model.DefaultLocation
+import com.github.odaridavid.weatherapp.model.Result
+import com.github.odaridavid.weatherapp.model.Weather
 
 interface WeatherRepository {
