+++ app/src/main/java/com/github/odaridavid/weatherapp/data/weather/DefaultWeatherRepository.kt
@@ -1,15 +1,15 @@
 package com.github.odaridavid.weatherapp.data.weather
 
-import com.github.odaridavid.weatherapp.core.Result
-import com.github.odaridavid.weatherapp.core.Result.Error
 import com.github.odaridavid.weatherapp.core.api.Logger
 import com.github.odaridavid.weatherapp.core.api.SettingsRepository
 import com.github.odaridavid.weatherapp.core.api.WeatherRepository
-import com.github.odaridavid.weatherapp.core.model.DefaultLocation
-import com.github.odaridavid.weatherapp.core.model.ExcludedData
-import com.github.odaridavid.weatherapp.core.model.Weather
 import com.github.odaridavid.weatherapp.data.weather.remote.RemoteWeatherDataSource
 import com.github.odaridavid.weatherapp.data.weather.remote.mapThrowableToErrorType
+import com.github.odaridavid.weatherapp.model.DefaultLocation
+import com.github.odaridavid.weatherapp.model.ExcludedData
+import com.github.odaridavid.weatherapp.model.Result
+import com.github.odaridavid.weatherapp.model.Result.Error
+import com.github.odaridavid.weatherapp.model.Weather
 import kotlinx.coroutines.flow.first
 import javax.inject.Inject
