+++ app/src/main/java/com/github/odaridavid/weatherapp/data/weather/remote/DefaultRemoteWeatherDataSource.kt
@@ -1,9 +1,9 @@
 package com.github.odaridavid.weatherapp.data.weather.remote
 
 import com.github.odaridavid.weatherapp.BuildConfig
-import com.github.odaridavid.weatherapp.core.Result
-import com.github.odaridavid.weatherapp.core.model.DefaultLocation
-import com.github.odaridavid.weatherapp.core.model.Weather
+import com.github.odaridavid.weatherapp.model.DefaultLocation
+import com.github.odaridavid.weatherapp.model.Result
+import com.github.odaridavid.weatherapp.model.Weather
 import javax.inject.Inject
 
 class DefaultRemoteWeatherDataSource @Inject constructor(
