+++ app/src/main/java/com/github/odaridavid/weatherapp/data/weather/remote/Mappers.kt
@@ -1,18 +1,18 @@
 package com.github.odaridavid.weatherapp.data.weather.remote
 
 import com.github.odaridavid.weatherapp.BuildConfig
-import com.github.odaridavid.weatherapp.core.ErrorType
-import com.github.odaridavid.weatherapp.core.model.ClientException
-import com.github.odaridavid.weatherapp.core.model.CurrentWeather
-import com.github.odaridavid.weatherapp.core.model.DailyWeather
-import com.github.odaridavid.weatherapp.core.model.GenericException
-import com.github.odaridavid.weatherapp.core.model.HourlyWeather
-import com.github.odaridavid.weatherapp.core.model.ServerException
-import com.github.odaridavid.weatherapp.core.model.Temperature
-import com.github.odaridavid.weatherapp.core.model.TimeFormat
-import com.github.odaridavid.weatherapp.core.model.Units
-import com.github.odaridavid.weatherapp.core.model.Weather
-import com.github.odaridavid.weatherapp.core.model.WeatherInfo
+import com.github.odaridavid.weatherapp.model.ClientException
+import com.github.odaridavid.weatherapp.model.CurrentWeather
+import com.github.odaridavid.weatherapp.model.DailyWeather
+import com.github.odaridavid.weatherapp.model.ErrorType
+import com.github.odaridavid.weatherapp.model.GenericException
+import com.github.odaridavid.weatherapp.model.HourlyWeather
+import com.github.odaridavid.weatherapp.model.ServerException
+import com.github.odaridavid.weatherapp.model.Temperature
+import com.github.odaridavid.weatherapp.model.TimeFormat
+import com.github.odaridavid.weatherapp.model.Units
+import com.github.odaridavid.weatherapp.model.Weather
+import com.github.odaridavid.weatherapp.model.WeatherInfo
 import java.io.IOException
 import java.net.HttpURLConnection
 import java.text.SimpleDateFormat
