+++ app/src/main/java/com/github/odaridavid/weatherapp/ui/settings/SettingsScreen.kt
@@ -11,16 +11,16 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.github.odaridavid.weatherapp.R
-import com.github.odaridavid.weatherapp.core.model.ExcludedData
-import com.github.odaridavid.weatherapp.core.model.SupportedLanguage
-import com.github.odaridavid.weatherapp.core.model.TimeFormat
-import com.github.odaridavid.weatherapp.core.model.Units
 import com.github.odaridavid.weatherapp.designsystem.WeatherAppTheme
 import com.github.odaridavid.weatherapp.designsystem.organism.MultiSelectBottomSheet
 import com.github.odaridavid.weatherapp.designsystem.organism.SettingOptionRow
 import com.github.odaridavid.weatherapp.designsystem.organism.SingleSelectBottomSheet
 import com.github.odaridavid.weatherapp.designsystem.organism.TopNavigationBar
 import com.github.odaridavid.weatherapp.designsystem.organism.VersionInfoText
+import com.github.odaridavid.weatherapp.model.ExcludedData
+import com.github.odaridavid.weatherapp.model.SupportedLanguage
+import com.github.odaridavid.weatherapp.model.TimeFormat
+import com.github.odaridavid.weatherapp.model.Units
 import kotlinx.coroutines.launch
 
 @OptIn(ExperimentalMaterial3Api::class)
