+++ app/src/main/java/com/github/odaridavid/weatherapp/ui/settings/UIMapper.kt
@@ -1,10 +1,10 @@
 package com.github.odaridavid.weatherapp.ui.settings
 
-import com.github.odaridavid.weatherapp.core.model.ExcludedData
-import com.github.odaridavid.weatherapp.core.model.SupportedLanguage
-import com.github.odaridavid.weatherapp.core.model.TimeFormat
-import com.github.odaridavid.weatherapp.core.model.Units
 import com.github.odaridavid.weatherapp.designsystem.organism.BottomSheetItem
+import com.github.odaridavid.weatherapp.model.ExcludedData
+import com.github.odaridavid.weatherapp.model.SupportedLanguage
+import com.github.odaridavid.weatherapp.model.TimeFormat
+import com.github.odaridavid.weatherapp.model.Units
 
 // TODO Fix repetition of mapping
 fun ExcludedData.toBottomSheetModel(isSelected: Boolean = false): BottomSheetItem {
