+++ shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/model/DefaultLocation.kt
@@ -1,3 +1,3 @@
-package com.github.odaridavid.weatherapp.core.model
+package com.github.odaridavid.weatherapp.model
 
 data class DefaultLocation(val longitude: Double, val latitude: Double)
