+++ shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/model/SupportedLanguage.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.core.model
+package com.github.odaridavid.weatherapp.model
 
 enum class SupportedLanguage(val languageName: String, val languageValue: String) {
     AFRIKAANS("Afrikaans", "af"),
