+++ shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/model/TimeFormat.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.core.model
+package com.github.odaridavid.weatherapp.model
 
 enum class TimeFormat(val value: String) {
     TWENTY_FOUR_HOUR("24 hours"),
