+++ shared/src/commonMain/kotlin/com/github/odaridavid/weatherapp/model/Units.kt
@@ -1,4 +1,4 @@
-package com.github.odaridavid.weatherapp.core.model
+package com.github.odaridavid.weatherapp.model
 
 enum class Units(val value: String, val tempLabel: String) {
     STANDARD("standard","°F"),
